Switch from Irmin 0.10 format to Git format
This will allow upgrading to newer versions of Irmin, which can't read
the old 0.10 format. If irmin-indexeddb detects an old-format store,
it will migrate the data to the new format. It puts the new data in
new `ao_git` and `rw_git` stores, but leaves the old data in `ao` and
`rw` in case the migration goes wrong.